Revert "fix: adjust markup and styling for hero section"
