+++ src/components/layout.js
@@ -6,10 +6,10 @@ export default function Layout({ children }) {
 	return (
 		<>
 			<header>
-				<nav className="nav-wrapper">
+				<nav id="navBar">
     
 					<div id="navigation">
-						<Link className="navLink" href="#hero" onClick={ e => scrollToElement( e ) }><p>home</p></Link>
+						<Link className="navLink" href="#home" onClick={ e => scrollToElement( e ) }><p>home</p></Link>
 						<Link className="navLink" href="#case" onClick={ e => scrollToElement( e ) }><p>projects</p></Link>
 						<Link className="navLink" href="#about" onClick={ e => scrollToElement( e ) }><p>about</p></Link>
 						<Link className="navLink" href="#connect" onClick={ e => scrollToElement( e ) }><p>connect</p></Link>
@@ -19,10 +19,11 @@ export default function Layout({ children }) {
 						<a href="mailto:dorothytoth@me.com"><img src="images/DT-comm-email-gray.svg" /></a>
 						<a href="https://www.linkedin.com/in/dorothy-toth/" target="_blank"><img src="images/DT-comm-linkedin-gray.svg" /></a>
 					</div>
-
-					<Link className="navLink" href="#hero" onClick={ e => scrollToElement( e ) }><img id="upToTop" src="images/DT-carrot-grayUp.svg" /></Link>
 					
 				</nav>
+
+				<Link className="navLink" href="#home" onClick={ e => scrollToElement( e ) }><img id="upToTop" src="images/DT-carrot-grayUp.svg" /></Link>
+
 			</header>
 			{ children }
 			<div className="footer"></div>
