+++ src/pages/index.js
@@ -1,4 +1,4 @@
-import HomeSection from "@/components/hero-section";
+import HomeSection from "@/components/home-section";
 import CaseStudies from "@/components/case-studies";
 import About from "@/components/about";
 import Connect from "@/components/connect";
