--- src/styles/hero.scss
@@ -1,87 +1,0 @@
-.section-hero {
-	margin: 0px;
-	display: flex;
-	position: relative;
-	min-height: 500px;
-	height: 70vh;
-	transition: 0.3s;
-
-	video {
-		mix-blend-mode: multiply;
-		object-fit: cover;
-		position: absolute;
-		top: 0;
-		left: 0;
-		height: 100%;
-		width: 100%;		
-	}
-
-	.hero-content {
-		position: absolute;
-		z-index: 2;
-		background: white;
-		border-radius: 10px;
-		background-color: rgba(255,255,255,.9);
-		right: 400px;
-		top: 160px;
-		padding: 50px;
-		box-shadow: 0 0 10px rgba(99,99,99,.1);
-		transition: .3s;
-		p {
-			font-size: 18px;
-			font-weight: 500;
-			margin: 0px;
-			color: #636364;
-		}
-	}
-
-	.hero-video-bg,
-	.hero-cta-wrapper,
-	.cta-image,
-	.cta-link {
-		background-position: 50%;
-		background-repeat: no-repeat;
-		background-size: cover;
-		align-items: center;
-		display: flex;
-		position: relative;
-		justify-content: flex-start;
-		transition: .3s;
-	}
-
-	.hero-video-bg {
-		flex: 0 0 70%;
-		background-color: #a3a8bf;
-	}
-
-	.hero-cta-wrapper {
-		flex: 0 0 30%;
-		display: flex;
-		flex-direction: column;
-		.cta-image {
-			flex: 1 0 200px;
-			width: 100%;
-			background-image: url(/images/screens.jpg);
-			background-color: #3670959b;
-			background-blend-mode: multiply;
-		}
-		.cta-link {
-			flex: 0 0 75px;
-			width: 100%;
-			height: 75px;
-			background-image: linear-gradient(70deg,#367095,#a3a8bf);
-			justify-content: flex-start;
-			cursor: pointer;
-			transition: .3s;
-			color: white;
-			font-size: 1.7rem;
-			padding: 0 1rem;
-			svg {
-				height: 2rem;
-				margin: 0 1rem;
-				fill: white;
-			}
-		}
-	}
-
-}
