--- src/styles/nav.scss
@@ -1,83 +1,0 @@
-.nav-wrapper {
-	display: flex;
-	position: fixed;
-	top: 0;
-	width: 100%;
-	flex-direction: row;
-	margin: 0px;
-	height: 75px;
-	z-index: 3;
-	background-color: rgba(255, 255, 255, 0.9);
-	align-items: center;
-	padding-right: 15px;
-	justify-content: space-between;
-	box-shadow: 0px 0px 10px rgba(99, 99, 99, 0.1);
-
-    #navigation {
-        display: flex;
-        width: 75%;
-        align-items: center;
-        justify-content: space-evenly;
-    }
-
-    #navigation p {
-        font-family: 'Kanit', sans-serif;
-        font-weight: 500;
-        color: #636364;
-        transition: 0.3s;
-    }
-
-    #navigation a:visited,
-    a:link {
-        cursor: pointer;
-        text-decoration: none;
-        color: #636364
-    }
-
-    #navigation p:hover {
-        cursor: pointer;
-        text-decoration: none;
-        color: #367095;
-    }
-
-    .icons {
-        display: flex;
-        justify-content: flex-end;
-        align-items: center;
-        width: 25%;
-        margin-right: 5%;
-    }
-
-    .icons img {
-        margin: 15px;
-        width: 30px;
-        cursor: pointer;
-        opacity: 0.5;
-        transition: 0.3s
-    }
-
-    .icons img:hover {
-        opacity: .9;
-    }
-
-    .carrot {
-        width: 30px;
-        cursor: pointer;
-    }
-
-    #upToTop {
-        width: 30px;
-        cursor: pointer;
-        position: fixed;
-        z-index: 99;
-        bottom: 20px;
-        right: 20px;
-        opacity: 0.5;
-        transition: 0.3s
-    }
-
-    #upToTop:hover {
-        opacity: .9;
-    }
-
-}
